# starter-resource-server/src/generated/swagger.yml
swagger: "2.0"
info:
    version: "0.0.1"
    title: "Swagger API"
host: localhost:4266
basePath: "/"
schemes:
    - http
    - https
consumes:
    - application/json
produces:
    - application/json
definitions:
...
    StoreCreateBody:
        type: object
        properties:
            name:
                type: string
...
paths:
...
    /users/{id}/store:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /stores:
        get:
            description: "sample description"
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
        post:
            description: "sample description"
            parameters:
                - name: body
                  description: "Not provided"
                  in: body
                  schema:
                    $ref: '#/definitions/StoreCreateBody'
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /stores/{id}:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /stores/{id}/videos:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
...
    /videos/{id}/store:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
...